Moving logic to apiserver to avoid dealing with session directly if some data sources are using it.
