+++ api/imagemetadata/client.go
@@ -57,3 +57,11 @@ func (c *Client) Save(metadata []params.CloudImageMetadata) ([]params.ErrorResul
 	}
 	return out.Results, nil
 }
+
+// UpdateFromPublishedImages lists currently published image metadata and
+// updates stored ones accordingly.
+// This method is primarily intended for a worker.
+func (c *Client) UpdateFromPublishedImages() error {
+	return errors.Trace(
+		c.facade.FacadeCall("UpdateFromPublishedImages", nil, nil))
+}
