+++ apiserver/imagemetadata/metadata.go
@@ -7,11 +7,20 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/juju/loggo"
+
+	"github.com/juju/errors"
 	"github.com/juju/juju/apiserver/common"
 	"github.com/juju/juju/apiserver/params"
+	"github.com/juju/juju/environs"
+	envmetadata "github.com/juju/juju/environs/imagemetadata"
+	"github.com/juju/juju/environs/simplestreams"
+	"github.com/juju/juju/juju/series"
 	"github.com/juju/juju/state"
 	"github.com/juju/juju/state/cloudimagemetadata"
 )
+
+var logger = loggo.GetLogger("juju.apiserver.imagemetadata")
 
 func init() {
 	common.RegisterStandardFacade("ImageMetadata", 1, NewAPI)
@@ -138,4 +147,100 @@ func parseMetadataFromParams(p params.CloudImageMetadata) cloudimagemetadata.Met
 		result.Stream = "released"
 	}
 	return result
+}
+
+// UpdateFromPublishedImages lists currently published image metadata and
+// updates stored ones accordingly.
+func (api *API) UpdateFromPublishedImages() error {
+	published, err := api.listPublished()
+	if err != nil {
+		return errors.Annotatef(err, "getting published images metadata")
+	}
+	err = api.saveAll(published)
+	return errors.Annotatef(err, "saving published images metadata")
+}
+
+func (api *API) listPublished() ([]*envmetadata.ImageMetadata, error) {
+	// Get environ
+	envCfg, err := api.metadata.EnvironConfig()
+	env, err := environs.New(envCfg)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+
+	// Get all images metadata sources for this environ.
+	sources, err := environs.ImageMetadataSources(env)
+	if err != nil {
+		return nil, err
+	}
+
+	// We want all metadata.
+	cons := envmetadata.NewImageConstraint(simplestreams.LookupParams{})
+	metadata, _, err := envmetadata.Fetch(sources, cons, false)
+	if err != nil {
+		return nil, err
+	}
+	return metadata, nil
+}
+
+func (api *API) saveAll(published []*envmetadata.ImageMetadata) error {
+	// Store converted metadata.Note that whether the metadata actually needs
+	// to be stored will be determined within this call.
+	errs, err := api.Save(convertToParams(published))
+	if err != nil {
+		return errors.Annotatef(err, "saving published images metadata")
+	}
+	return processErrors(errs.Results)
+}
+
+// convertToParams converts environment-specific images metadata to structured metadata format.
+var convertToParams = func(published []*envmetadata.ImageMetadata) params.MetadataSaveParams {
+	metadata := make([]params.CloudImageMetadata, len(published))
+	for i, p := range published {
+		metadata[i] = params.CloudImageMetadata{
+			Source:          "public",
+			ImageId:         p.Id,
+			Stream:          p.Stream,
+			Region:          p.RegionName,
+			Arch:            p.Arch,
+			VirtualType:     p.VirtType,
+			RootStorageType: p.Storage,
+		}
+		// Translate version (eg.14.04) to a series (eg. "trusty")
+		metadata[i].Series = versionSeries(p.Version)
+	}
+
+	return params.MetadataSaveParams{Metadata: metadata}
+}
+
+// TODO(dfc) why is this like this ?
+var seriesVersion = series.SeriesVersion
+
+func versionSeries(v string) string {
+	if v == "" {
+		return v
+	}
+	for _, s := range series.SupportedSeries() {
+		sv, err := seriesVersion(s)
+		if err != nil {
+			logger.Errorf("cannot determine version for series %v: %v", s, err)
+		}
+		if v == sv {
+			return s
+		}
+	}
+	return v
+}
+
+func processErrors(errs []params.ErrorResult) error {
+	msgs := []string{}
+	for _, e := range errs {
+		if e.Error != nil && e.Error.Message != "" {
+			msgs = append(msgs, e.Error.Message)
+		}
+	}
+	if len(msgs) != 0 {
+		return errors.Errorf("saving some image metadata:\n%v", strings.Join(msgs, "\n"))
+	}
+	return nil
 }
